compile definition of `vol::win::Com` separately
