+++ core/windows/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library(volume_core_windows INTERFACE)
-target_include_directories(volume_core_windows INTERFACE include)
+add_library(volume_core_windows STATIC src/com.cpp)
+target_include_directories(volume_core_windows PUBLIC include)
